hack to reduce length of the encryption of userIDs to 10 in order to use them for destination tags.

+++ src/test/java/unit/TestHastyPuddingCipherUtil.java
@@ -24,7 +24,7 @@ public class TestHastyPuddingCipherUtil
 	{
 		long beforeEncrypt = 1234567890l;
 		//BigInteger[] kx = HastyPuddingCipherUtil.createKeyExpansionTable(2, 64);
-		BigInteger encrypted = HastyPuddingCipherUtil.encryptHPCShort(beforeEncrypt);
+		Long encrypted = HastyPuddingCipherUtil.encryptHPCShort(beforeEncrypt);
 		assertFalse(beforeEncrypt == encrypted.longValue());
 		Long decrypted = HastyPuddingCipherUtil.decryptHPCShort(encrypted);
 		assertTrue(beforeEncrypt == decrypted);
